requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pgvector==0.2.4

# AWS
boto3==1.29.7
botocore==1.32.7

# ML and embeddings
sentence-transformers==2.2.2
pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
numpy==1.24.3

# Dashboard
dash==2.14.2
plotly==5.18.0
dash-bootstrap-components==1.5.0
pandas==2.1.4

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
